ignore npm errors with --production set
In (and only in) --production, we use `npm ls`. If `npm ls` encounters a
dependency graph integrity issue (such as an unmet peer dependency), it
still produces valid JSON (to stdout), but exits with a non-zero code.
This meant we were checking for dependency graph integrity with the
--production flag set.

|                    | always                          |
| ------------------ | ------------------------------- |
| --production true  | fail on dependency graph errors |
| --production false | ignore dependency graph errors  |

With this change, we ignore npm's exit code. If we are able to parse the
json it outputs, we will analyze the dependencies. If we are not, we
will exit later with `cannot interpret npm ls --json output`.