gitignore /.nyc-output, /node_modules, /package-lock.json
These files materialize after running `npm install` and `npm run test`.
Adding them to `.gitignore` to reduce friction for developers and to
help ensure that they are not accidentally committed.

I found this discussion from a few years ago that seems relevant:

https://github.com/jslicense/licensee.js/pull/58#issuecomment-558713096

While I agree with the sentiment of something so universal being a
ignored by default globally, since it is not the default git behavior in
practice it makes it more difficult for people to contribute to this
repo. The cost of adding these lines to `.gitignore` is low, so this
seems like a reasonable change to make.

I was surprised that this was not already the case, and since this seems
to be a common issue in this repo, it seems that other developers are
also surprised by this.

- https://github.com/jslicense/licensee.js/pull/23
- https://github.com/jslicense/licensee.js/pull/26