Fix test with unlicensed subdependency
This test broke because the order of output changed, not because
Licensee produced the wrong results.

Previously, the test checked for exact output, line by line.  This PR
adds the `--ndjson` flag to the invocation, parses the NDJSON output,
and tests the results logically, without requiring package analyses be
output in any particular order.